refactor(utils): move error handling functions to errors.go
- Move IgnoreAlreadyExists, IgnoreNotFound, and IgnoreNoKindMatchError functions from crtl_utils.go to errors.go
