+++ pkg/utils/errors.go
@@ -19,8 +19,32 @@ package utils
 import (
 	"github.com/go-logr/logr"
 	apierrs "k8s.io/apimachinery/pkg/api/errors"
+	apimeta "k8s.io/apimachinery/pkg/api/meta"
 	"k8s.io/apimachinery/pkg/types"
 )
+
+// IgnoreAlreadyExists ignores already existes error
+func IgnoreAlreadyExists(err error) error {
+	if apierrs.IsAlreadyExists(err) {
+		return nil
+	}
+	return err
+}
+
+// IgnoreNotFound ignores not found
+func IgnoreNotFound(err error) error {
+	if apierrs.IsNotFound(err) {
+		return nil
+	}
+	return err
+}
+
+func IgnoreNoKindMatchError(err error) error {
+	if apimeta.IsNoMatchError(err) {
+		return nil
+	}
+	return err
+}
 
 // LoggingErrorExceptConflict logs error except for updating operation violates with etcd concurrency control
 func LoggingErrorExceptConflict(logging logr.Logger, err error, info string, namespacedKey types.NamespacedName) (result error) {
